[update]ログインユーザー以外の編集、削除を行えないように変更

+++ app/controllers/public/cart_products_controller.rb
@@ -1,5 +1,6 @@
 class Public::CartProductsController < ApplicationController
   def index
+    
   end
 
   def create
@@ -14,25 +15,29 @@ class Public::CartProductsController < ApplicationController
 
   def update
     cart_product = CartProduct.find(params[:id])
-    cart_product.update(cart_product_params)
-    redirect_back(fallback_location: root_path)
+    if cart_product.customer_id == current_customer.id
+      cart_product.update(cart_product_params)
+    end
+    redirect_to cart_products_path
   end
 
   def destroy
     cart_product = CartProduct.find(params[:id])
-    cart_product.destroy
-    redirect_back(fallback_location: root_path)
+    if cart_product.customer_id == current_customer.id
+      cart_product.destroy
+    end
+    redirect_to cart_products_path
   end
 
   def destroy_all
     cart_products = CartProduct.where(customer_id: current_customer.id)
     cart_products.destroy_all
-    redirect_back(fallback_location: root_path)
+    redirect_to cart_products_path
   end
-  
+
   private
-  
+
   def cart_product_params
-    params.require(:cart_product).permit(:product_id, :quantity)	
+    params.require(:cart_product).permit(:product_id, :quantity)
   end
 end
